requirements.txt
asgiref==3.7.2
crispy-tailwind==0.5.0
Django==4.2.8
django-crispy-forms==2.1
django-extensions==3.2.3
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
